Some adjustements done

+++ backend/src/main/java/es/udc/asiproject/service/impl/SaleServiceImpl.java
@@ -1,8 +1,5 @@
 package es.udc.asiproject.service.impl;
 
-import java.io.FileNotFoundException;
-import java.io.FileOutputStream;
-import java.io.PrintWriter;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,12 +9,8 @@ import org.springframework.stereotype.Service;
 
 import es.udc.asiproject.persistence.dao.SaleDao;
 import es.udc.asiproject.persistence.dao.UserDao;
-import es.udc.asiproject.persistence.model.Accommodation;
-import es.udc.asiproject.persistence.model.Activity;
 import es.udc.asiproject.persistence.model.Sale;
 import es.udc.asiproject.persistence.model.Sale.SaleState;
-import es.udc.asiproject.persistence.model.Transport;
-import es.udc.asiproject.persistence.model.Travel;
 import es.udc.asiproject.persistence.model.User;
 import es.udc.asiproject.persistence.model.User.RoleType;
 import es.udc.asiproject.service.SaleService;
@@ -78,7 +71,7 @@ public class SaleServiceImpl implements SaleService {
 
 		if (saleToDelete.isPresent()) {
 
-			if (saleToDelete.get().getAgent().getId() == userId) {
+			if (saleToDelete.get().getAgent().getId() == userId && saleToDelete.get().getState() == SaleState.NORMAL) {
 
 				saleDao.delete(saleToDelete.get());
 
@@ -94,10 +87,11 @@ public class SaleServiceImpl implements SaleService {
 	public void paySale(Long userId, Long saleId) throws InstanceNotFoundException, PermissionException {
 
 		Optional<Sale> saleToPay = saleDao.findById(saleId);
+		Optional<User> userToPay = userDao.findById(userId);
 
-		if (saleToPay.isPresent()) {
+		if (saleToPay.isPresent() && userToPay.isPresent()) {
 
-			if (saleToPay.get().getClient().getId() == userId) {
+			if (saleToPay.get().getClient().getId() == userId || userToPay.get().getRole() == RoleType.GERENTE) {
 
 				saleToPay.get().setState(SaleState.PAID);
 				saleDao.save(saleToPay.get());
@@ -110,63 +104,69 @@ public class SaleServiceImpl implements SaleService {
 
 	}
 
-	@Override
-	public void generateBill(Long userId, Long saleId)
-			throws InstanceNotFoundException, PermissionException, FileNotFoundException {
-
-		Optional<Sale> OptionalSaleToGenerateBill = saleDao.findById(saleId);
-
-		Optional<User> OptionalUser = userDao.findById(userId);
-
-		if (OptionalSaleToGenerateBill.isPresent() && OptionalUser.isPresent()
-				&& OptionalSaleToGenerateBill.get().getState() == SaleState.PAID) {
-
-			Sale saleToGenerateBill = OptionalSaleToGenerateBill.get();
-			User user = OptionalUser.get();
-
-			if (user.getId() == saleToGenerateBill.getAgent().getId()
-					|| user.getId() == saleToGenerateBill.getClient().getId() || user.getRole() == RoleType.GERENTE) {
-
-				String ruta = "C:\\Users\\jaime\\Desktop\\Universidad\\MUEI\\ASI\\Prácticas\\Bills\bill.txt";
-
-				PrintWriter file = new PrintWriter((new FileOutputStream(ruta, false)));
-
-				file.println("Estado: " + saleToGenerateBill.getState());
-				file.println("Precio: " + saleToGenerateBill.getPrice());
-				file.println("Fecha: " + saleToGenerateBill.getCreatedAt());
-				file.println("Agente: " + saleToGenerateBill.getAgent().getFirstName() + " "
-						+ saleToGenerateBill.getAgent().getLastName());
-				file.println("Cliente: " + saleToGenerateBill.getClient().getFirstName() + " "
-						+ saleToGenerateBill.getClient().getLastName());
-				file.println("");
-				file.println("Alojamientos incluídos: ");
-				for (Accommodation accommodation : saleToGenerateBill.getAccommodations()) {
-
-					file.println(accommodation.getName());
-				}
-				file.println("");
-				file.println("Actividades incluídas: ");
-				for (Activity activity : saleToGenerateBill.getActivities()) {
-
-					file.println(activity.getName());
-				}
-				file.println("");
-				file.println("Transportes incluídos: ");
-				for (Transport transport : saleToGenerateBill.getTransports()) {
-
-					file.println(transport.getName());
-				}
-				file.println("");
-				file.println("Viajes incluídos: ");
-				for (Travel travel : saleToGenerateBill.getTravels()) {
-
-					file.println(travel.getName());
-				}
-
-			}
-
-		} else
-			throw new InstanceNotFoundException(Sale.class.getSimpleName(), saleId);
-	}
+//	@Override
+//	public ByteArrayResource generateBill(Long userId, Long saleId)
+//			throws InstanceNotFoundException, PermissionException, FileNotFoundException, IOException {
+//
+//		Optional<Sale> OptionalSaleToGenerateBill = saleDao.findById(saleId);
+//
+//		Optional<User> OptionalUser = userDao.findById(userId);
+//
+//		if (OptionalSaleToGenerateBill.isPresent() && OptionalUser.isPresent()
+//				&& OptionalSaleToGenerateBill.get().getState() == SaleState.PAID) {
+//
+//			Sale saleToGenerateBill = OptionalSaleToGenerateBill.get();
+//			User user = OptionalUser.get();
+//
+//			if (user.getId() == saleToGenerateBill.getAgent().getId()
+//					|| user.getId() == saleToGenerateBill.getClient().getId() || user.getRole() == RoleType.GERENTE) {
+//
+//				String ruta = "../src/main/resources/bill.txt";
+//
+//				PrintWriter file = new PrintWriter((new FileOutputStream(ruta, false)));
+//
+//				file.println("Estado: " + saleToGenerateBill.getState());
+//				file.println("Precio: " + saleToGenerateBill.getPrice());
+//				file.println("Fecha: " + saleToGenerateBill.getCreatedAt());
+//				file.println("Agente: " + saleToGenerateBill.getAgent().getFirstName() + " "
+//						+ saleToGenerateBill.getAgent().getLastName());
+//				file.println("Cliente: " + saleToGenerateBill.getClient().getFirstName() + " "
+//						+ saleToGenerateBill.getClient().getLastName());
+//				file.println("");
+//				file.println("Alojamientos incluídos: ");
+//				for (Accommodation accommodation : saleToGenerateBill.getAccommodations()) {
+//
+//					file.println(accommodation.getName());
+//				}
+//				file.println("");
+//				file.println("Actividades incluídas: ");
+//				for (Activity activity : saleToGenerateBill.getActivities()) {
+//
+//					file.println(activity.getName());
+//				}
+//				file.println("");
+//				file.println("Transportes incluídos: ");
+//				for (Transport transport : saleToGenerateBill.getTransports()) {
+//
+//					file.println(transport.getName());
+//				}
+//				file.println("");
+//				file.println("Viajes incluídos: ");
+//				for (Travel travel : saleToGenerateBill.getTravels()) {
+//
+//					file.println(travel.getName());
+//				}
+//
+//				Path path = Paths.get(ruta);
+//				ByteArrayResource resource = new ByteArrayResource(Files.readAllBytes(path));
+//
+//				return resource;
+//			}
+//
+//		} else
+//			throw new InstanceNotFoundException(Sale.class.getSimpleName(), saleId);
+//
+//		return null;
+//	}
 
 }
